+++ src/consts/options.js
@@ -36,3 +36,8 @@ export const PRICE_HISTORY_GRANULARITY_OPTIONS = {
   day: { title: 'Day', isDefault: true },
   hour: { title: 'Hour' },
 };
+
+export const GAS_GRANULARITY_OPTIONS = {
+  day: { title: 'Day', isDefault: true },
+  hour: { title: 'Hour' },
+};