+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug fixes
 
 ### Features
+- Add message type gas statistics #260
 - Add calls to getAssetMetadata where explicitly needed #283
 
 ## 2.6.0